requirements.txt
Flask>=3.0.0
PyMySQL>=1.1.0
Werkzeug>=3.0.0
gunicorn>=21.0.0
Pillow>=10.0.0
